# packages/api-client/src/generated/api.schemas.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * Roastline coffee shop operations API
 * OpenAPI spec version: 0.2.0
 */
export interface HealthStatus {
  status: string;
}

export interface RegisterInput {
  name: string;
  email: string;
  password: string;
}

export interface LoginInput {
  email: string;
  password: string;
}

export interface StaffAccount {
  id: string;
  name: string;
  email: string;
  role: string;
  color: string;
  createdAt: string;
}

export interface SalesPoint {
  label: string;
  value: number;
}

export interface Activity {
  id: string;
  title: string;
  detail: string;
  time: string;
  tone: string;
}

export interface Dashboard {
  todaySalesCents: number;
  salesChangePercent: number;
  ordersToday: number;
  averageTicketCents: number;
  laborPercent: number;
  lowStockCount: number;
  openTickets: number;
  salesByHour: SalesPoint[];
  recentActivity: Activity[];
}

export interface MenuCategory {
  id: string;
  name: string;
  sortOrder: number;
}

export interface MenuItem {
  id: string;
  categoryId: string;
  name: string;
  description: string;
  priceCents: number;
  available: boolean;
  bestseller: boolean;
  color: string;
  imageUrl?: string | null;
  recipe: string[];
}

export interface Modifier {
  id: string;
  name: string;
  priceDeltaCents: number;
}

export interface ModifierGroup {
  id: string;
  name: string;
  required: boolean;
  options: Modifier[];
}

export interface MenuResponse {
  categories: MenuCategory[];
  items: MenuItem[];
  modifierGroups: ModifierGroup[];
}

export interface MenuItemInput {
  categoryId: string;
  name: string;
  description: string;
  priceCents: number;
  imageUrl?: string;
}

export interface AvailabilityInput {
  available: boolean;
}

export interface InventoryItem {
  id: string;
  name: string;
  unit: string;
  onHand: number;
  parLevel: number;
  costPerUnitCents: number;
  status: string;
  lastCounted: string;
}

export interface InventoryCountInput {
  onHand: number;
}

export interface Order {
  id: string;
  clientGeneratedUuid: string;
  channel: string;
  status: string;
  itemSummary: string;
  totalCents: number;
  paymentMethod: string;
  createdAt: string;
  customerName: string;
}

export interface OrderItemInput {
  menuItemId: string;
  quantity: number;
}

export interface OrderInput {
  clientGeneratedUuid: string;
  channel: string;
  paymentMethod: string;
  customerName?: string;
  items: OrderItemInput[];
}

export interface OrderStatusInput {
  status: string;
}

export interface StaffMember {
  id: string;
  name: string;
  role: string;
  initials: string;
  hoursThisWeek: number;
  status: string;
  color: string;
}

export interface Shift {
  id: string;
  staffId: string;
  staffName: string;
  day: string;
  start: string;
  end: string;
  status: string;
}

export interface ShiftInput {
  staffId: string;
  day: string;
  start: string;
  end: string;
}

export interface LoyaltyTier {
  name: string;
  points: number;
  reward: string;
  members: number;
}

export interface LoyaltySnapshot {
  members: number;
  activeThisMonth: number;
  pointsIssued: number;
  redemptionRate: number;
  tiers: LoyaltyTier[];
}

export interface SalesReport {
  periodLabel: string;
  totalSalesCents: number;
  grossSalesCents: number;
  netSalesCents: number;
  taxCents: number;
  tipsCents: number;
  averageTicketCents: number;
  daily: SalesPoint[];
}