# packages/api-schema/src/generated/types/dashboard.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * Roastline coffee shop operations API
 * OpenAPI spec version: 0.2.0
 */
import type { Activity } from './activity';
import type { SalesPoint } from './salesPoint';

export interface Dashboard {
  todaySalesCents: number;
  salesChangePercent: number;
  ordersToday: number;
  averageTicketCents: number;
  laborPercent: number;
  lowStockCount: number;
  openTickets: number;
  salesByHour: SalesPoint[];
  recentActivity: Activity[];
}