# packages/api-schema/src/generated/types/loginInput.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * Roastline coffee shop operations API
 * OpenAPI spec version: 0.2.0
 */

export interface LoginInput {
  email: string;
  password: string;
}