# packages/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.2.0
  description: Roastline coffee shop operations API
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
  - name: auth
  - name: dashboard
  - name: menu
  - name: inventory
  - name: orders
  - name: staff
  - name: loyalty
  - name: reports
paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"
  /auth/register:
    post:
      operationId: register
      tags: [auth]
      summary: Create the first (owner) account. Fails once any staff account exists.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterInput"
      responses:
        "201":
          description: Created and signed in
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StaffAccount"
  /auth/login:
    post:
      operationId: login
      tags: [auth]
      summary: Sign in with email and password
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginInput"
      responses:
        "200":
          description: Signed in
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StaffAccount"
  /auth/logout:
    post:
      operationId: logout
      tags: [auth]
      summary: End the current session
      responses:
        "204":
          description: Signed out
  /auth/me:
    get:
      operationId: getCurrentStaff
      tags: [auth]
      summary: Get the signed-in staff account
      responses:
        "200":
          description: Current staff account
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StaffAccount"
  /dashboard:
    get:
      operationId: getDashboard
      tags: [dashboard]
      summary: Get the current operating snapshot
      responses:
        "200":
          description: Dashboard snapshot
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Dashboard"
  /menu:
    get:
      operationId: getMenu
      tags: [menu]
      summary: Get menu categories and items
      responses:
        "200":
          description: Menu
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MenuResponse"
  /menu/items:
    post:
      operationId: createMenuItem
      tags: [menu]
      summary: Create a menu item
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MenuItemInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MenuItem"
  /menu/items/{id}/availability:
    patch:
      operationId: updateMenuItemAvailability
      tags: [menu]
      summary: Toggle an item availability state
      parameters:
        - $ref: "#/components/parameters/Id"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AvailabilityInput"
      responses:
        "200":
          description: Updated item
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MenuItem"
  /inventory:
    get:
      operationId: getInventory
      tags: [inventory]
      summary: Get ingredient stock levels
      responses:
        "200":
          description: Inventory
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/InventoryItem"
  /inventory/{id}/count:
    post:
      operationId: recordInventoryCount
      tags: [inventory]
      summary: Record a physical count
      parameters:
        - $ref: "#/components/parameters/Id"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/InventoryCountInput"
      responses:
        "200":
          description: Updated inventory item
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/InventoryItem"
  /orders:
    get:
      operationId: getOrders
      tags: [orders]
      summary: Get recent orders
      responses:
        "200":
          description: Recent orders
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Order"
    post:
      operationId: createOrder
      tags: [orders]
      summary: Create and complete a POS order
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/OrderInput"
      responses:
        "201":
          description: Completed order
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"
  /orders/{id}/status:
    patch:
      operationId: updateOrderStatus
      tags: [orders]
      summary: Move a prep ticket through its workflow
      parameters:
        - $ref: "#/components/parameters/Id"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/OrderStatusInput"
      responses:
        "200":
          description: Updated order
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"
  /staff:
    get:
      operationId: getStaff
      tags: [staff]
      summary: Get staff and schedule summary
      responses:
        "200":
          description: Staff
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/StaffMember"
  /shifts:
    get:
      operationId: getShifts
      tags: [staff]
      summary: Get published shifts
      responses:
        "200":
          description: Shifts
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Shift"
    post:
      operationId: createShift
      tags: [staff]
      summary: Add a shift
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ShiftInput"
      responses:
        "201":
          description: Created shift
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Shift"
  /loyalty:
    get:
      operationId: getLoyalty
      tags: [loyalty]
      summary: Get loyalty program snapshot
      responses:
        "200":
          description: Loyalty snapshot
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LoyaltySnapshot"
  /reports/sales:
    get:
      operationId: getSalesReport
      tags: [reports]
      summary: Get sales report
      responses:
        "200":
          description: Sales report
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SalesReport"
components:
  parameters:
    Id:
      name: id
      in: path
      required: true
      schema:
        type: string
  schemas:
    HealthStatus:
      type: object
      properties:
        status: { type: string }
      required: [status]
    RegisterInput:
      type: object
      required: [name, email, password]
      properties:
        name: { type: string }
        email: { type: string }
        password: { type: string }
    LoginInput:
      type: object
      required: [email, password]
      properties:
        email: { type: string }
        password: { type: string }
    StaffAccount:
      type: object
      required: [id, name, email, role, color, createdAt]
      properties:
        id: { type: string }
        name: { type: string }
        email: { type: string }
        role: { type: string }
        color: { type: string }
        createdAt: { type: string }
    Dashboard:
      type: object
      required: [todaySalesCents, salesChangePercent, ordersToday, averageTicketCents, laborPercent, lowStockCount, openTickets, salesByHour, recentActivity]
      properties:
        todaySalesCents: { type: number }
        salesChangePercent: { type: number }
        ordersToday: { type: number }
        averageTicketCents: { type: number }
        laborPercent: { type: number }
        lowStockCount: { type: number }
        openTickets: { type: number }
        salesByHour:
          type: array
          items: { $ref: "#/components/schemas/SalesPoint" }
        recentActivity:
          type: array
          items: { $ref: "#/components/schemas/Activity" }
    SalesPoint:
      type: object
      required: [label, value]
      properties:
        label: { type: string }
        value: { type: number }
    Activity:
      type: object
      required: [id, title, detail, time, tone]
      properties:
        id: { type: string }
        title: { type: string }
        detail: { type: string }
        time: { type: string }
        tone: { type: string }
    MenuResponse:
      type: object
      required: [categories, items, modifierGroups]
      properties:
        categories:
          type: array
          items: { $ref: "#/components/schemas/MenuCategory" }
        items:
          type: array
          items: { $ref: "#/components/schemas/MenuItem" }
        modifierGroups:
          type: array
          items: { $ref: "#/components/schemas/ModifierGroup" }
    MenuCategory:
      type: object
      required: [id, name, sortOrder]
      properties:
        id: { type: string }
        name: { type: string }
        sortOrder: { type: number }
    MenuItem:
      type: object
      required: [id, categoryId, name, description, priceCents, available, bestseller, color, recipe]
      properties:
        id: { type: string }
        categoryId: { type: string }
        name: { type: string }
        description: { type: string }
        priceCents: { type: number }
        available: { type: boolean }
        bestseller: { type: boolean }
        color: { type: string }
        imageUrl: { type: string, nullable: true }
        recipe:
          type: array
          items: { type: string }
    MenuItemInput:
      type: object
      required: [categoryId, name, description, priceCents]
      properties:
        categoryId: { type: string }
        name: { type: string }
        description: { type: string }
        priceCents: { type: number }
        imageUrl: { type: string }
    ModifierGroup:
      type: object
      required: [id, name, required, options]
      properties:
        id: { type: string }
        name: { type: string }
        required: { type: boolean }
        options:
          type: array
          items: { $ref: "#/components/schemas/Modifier" }
    Modifier:
      type: object
      required: [id, name, priceDeltaCents]
      properties:
        id: { type: string }
        name: { type: string }
        priceDeltaCents: { type: number }
    AvailabilityInput:
      type: object
      required: [available]
      properties:
        available: { type: boolean }
    InventoryItem:
      type: object
      required: [id, name, unit, onHand, parLevel, costPerUnitCents, status, lastCounted]
      properties:
        id: { type: string }
        name: { type: string }
        unit: { type: string }
        onHand: { type: number }
        parLevel: { type: number }
        costPerUnitCents: { type: number }
        status: { type: string }
        lastCounted: { type: string }
    InventoryCountInput:
      type: object
      required: [onHand]
      properties:
        onHand: { type: number }
    Order:
      type: object
      required: [id, clientGeneratedUuid, channel, status, itemSummary, totalCents, paymentMethod, createdAt, customerName]
      properties:
        id: { type: string }
        clientGeneratedUuid: { type: string }
        channel: { type: string }
        status: { type: string }
        itemSummary: { type: string }
        totalCents: { type: number }
        paymentMethod: { type: string }
        createdAt: { type: string }
        customerName: { type: string }
    OrderInput:
      type: object
      required: [clientGeneratedUuid, channel, paymentMethod, items]
      properties:
        clientGeneratedUuid: { type: string }
        channel: { type: string }
        paymentMethod: { type: string }
        customerName: { type: string }
        items:
          type: array
          items: { $ref: "#/components/schemas/OrderItemInput" }
    OrderItemInput:
      type: object
      required: [menuItemId, quantity]
      properties:
        menuItemId: { type: string }
        quantity: { type: number }
    OrderStatusInput:
      type: object
      required: [status]
      properties:
        status: { type: string }
    StaffMember:
      type: object
      required: [id, name, role, initials, hoursThisWeek, status, color]
      properties:
        id: { type: string }
        name: { type: string }
        role: { type: string }
        initials: { type: string }
        hoursThisWeek: { type: number }
        status: { type: string }
        color: { type: string }
    Shift:
      type: object
      required: [id, staffId, staffName, day, start, end, status]
      properties:
        id: { type: string }
        staffId: { type: string }
        staffName: { type: string }
        day: { type: string }
        start: { type: string }
        end: { type: string }
        status: { type: string }
    ShiftInput:
      type: object
      required: [staffId, day, start, end]
      properties:
        staffId: { type: string }
        day: { type: string }
        start: { type: string }
        end: { type: string }
    LoyaltySnapshot:
      type: object
      required: [members, activeThisMonth, pointsIssued, redemptionRate, tiers]
      properties:
        members: { type: number }
        activeThisMonth: { type: number }
        pointsIssued: { type: number }
        redemptionRate: { type: number }
        tiers:
          type: array
          items: { $ref: "#/components/schemas/LoyaltyTier" }
    LoyaltyTier:
      type: object
      required: [name, points, reward, members]
      properties:
        name: { type: string }
        points: { type: number }
        reward: { type: string }
        members: { type: number }
    SalesReport:
      type: object
      required: [periodLabel, totalSalesCents, grossSalesCents, netSalesCents, taxCents, tipsCents, averageTicketCents, daily]
      properties:
        periodLabel: { type: string }
        totalSalesCents: { type: number }
        grossSalesCents: { type: number }
        netSalesCents: { type: number }
        taxCents: { type: number }
        tipsCents: { type: number }
        averageTicketCents: { type: number }
        daily:
          type: array
          items: { $ref: "#/components/schemas/SalesPoint" }